requirements.txt
Django==1.7.4
SQLAlchemy==0.9.8
dj-database-url==0.3.0
django-postgrespool==0.3.0
django-filter==0.9.2
djangorestframework==3.1.0
Markdown==2.6.1
waitress==0.8.9
psycopg2==2.6
whitenoise==1.0.6
wsgiref==0.1.2
